% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{$<-.experiment}
\alias{$<-.experiment}
\title{Replace a column of an experiment}
\usage{
\method{$}{experiment}(exp, i) <- value
}
\arguments{
\item{exp}{An object of class \code{experiment}.}

\item{i}{A column index.}

\item{value}{A vector used to replace the values of the indexed column.}
}
\value{
An object of class \code{experiment}.
}
\description{
Replaces a column of an experiment with new value(s).
}
\details{
If the length of the vector used to replace the column is not the same as the
original number of rows of the experiment, there is duplication of the
shortest element.
}
\examples{
# Here is an experiment with 1 simulation:
sir1 <- load_experiment("sir", system.file("examples", "sir.gaml", package = "rama"), "sir")
sir1
# Let's replace the value of the "p_S0" column by a vector of 3 values:
sir2 <- sir1
sir2$p_S0 <- 1:3
# We can check that it automatically expands the number of simulations:
sir2
# If, on the contrary, we now replace the values of "p_S0" of "sir2" by a
# single value:
sir3 <- sir2
sir3$p_S0 <- 2
# We can check that it automatically reduces the number of simulations (if
# the replacement leads to an experiment with exactly identical simulations):
sir3
# If you wish to delete one column:
sir3$r_R <- NULL
sir3

}
