% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R, R/is.experiment.R
\name{experiment}
\alias{experiment}
\alias{is.experiment}
\title{Constructor of experiments}
\usage{
experiment(parameters, obsrates, tmax, seed, experiment, model, dir = "",
  dic = NULL)

is.experiment(exp)
}
\arguments{
\item{parameters}{A data frame of numerical values giving the values of each
parameter (in column), for each simulation (in row).}

\item{obsrates}{A data frame of positive integer values giving the periods,
in time steps, at which observed variables are observed.
Should have the same number of rows as \code{parameters}.}

\item{tmax}{A positive integer vector, the length of which is either 1 or
equal to the number of \code{parameters} and \code{obsrates}: it
gives the end of simulations, in numbers of time steps.}

\item{seed}{A numerical vector, the length of which is either 1 or equal to
the number of \code{parameters} and \code{obsrates}: it gives the
seeds of the simulations.}

\item{experiment}{The name of an experiment of the \code{GAML} file
\code{model}.}

\item{model}{The path to a \code{GAML} file.}

\item{dir}{The path to a directory where the simulations output are saved. If
not specified (i.e. empty character string, default), then the
simulations are saved in a subdirectory of the working directory,
the name of which is made from the name of \code{model}. See
\code{Details} for more information.}

\item{dic}{A named vector of character strings. The values and the names of
this vector should be consistent with the names of
\code{parameters}, \code{obsrates} as well as the variables and
parameters defined in the \code{model} \code{GAML} file. See
Details for more information.}

\item{exp}{object to be tested}
}
\value{
An object of class \code{experiment}.

The function returns `TRUE` or `FALSE` depending on whether its
argument is of chatacter type or not
}
\description{
Allows to build an object of class \code{experiment} from individual parts.
}
\details{
The class \code{experiment} inherits from the class \code{tibble}
(\code{tbl_df}). It contains parameters values as well as periods of
observation of the observed variables and it connects to a \code{GAML} file
that specifies the full model as well as to a directory that contains the
outputs of simuations.

TO DO: Explains how \code{dir} is made is not specified. Explains what
\code{dic} is.
}
\examples{
# to test if an object is a class `experiment`
df <- data.frame("S0" = rep(999, 5), "I0" = rep(1, 5), "R0" = rep(0, 5),
                "beta" = rep(1.5, 5), "gamma" = runif (5, 0, 1),
                "S" = rep(1, 5), "I" = rep(1, 5), "R" = rep(1, 5),
                "a" = rep(1000, 5), "b" = rep(1, 5))
exp <- as_experiment(df, parameters = c("S0", "I0", "R0", "beta", "gamma"),
                     obsrates = c("S", "I", "R"),
                     tmax = "a",
                     seed = "b",
                     experiment = "sir",
                     model =
                       system.file("examples", "sir.gaml", package = "rama"))
is.experiment(exp)
}
