% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{expname}
\alias{expname}
\alias{expname.default}
\alias{expname.experiment}
\title{Get the name of an experiment}
\usage{
expname(x)

\method{expname}{default}(x)

\method{expname}{experiment}(x)
}
\arguments{
\item{x}{An object of class \code{experiment}.}
}
\value{
The name of the experiment that the inputed \code{experiment} object
is linked to.
}
\description{
Retrieves the name of the experiment that an \code{experiment} object is
linked to.
}
\examples{
exp1 <- load_experiment("sir", system.file("examples", "sir.gaml", package = "rama"))
expname(exp1)

}
