% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gama_local_distrib_path}
\alias{gama_local_distrib_path}
\title{In function of remote distribution of the OS returns the path}
\usage{
gama_local_distrib_path()
}
\description{
In function of remote distribution of the OS returns the path
}
