% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gama_remote_distrib}
\alias{gama_remote_distrib}
\title{In function of remote distribution of the OS returns the path}
\usage{
gama_remote_distrib()
}
\description{
In function of remote distribution of the OS returns the path
}
