% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rama.R
\docType{package}
\name{gamar-package}
\alias{gamar-package}
\title{gamar: An interface to the GAMA simulation platform.}
\description{
The gamar package provides the function 'run' that can run a GAMA experiment
as described in an XML file.
}
