% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get information on the model}
\usage{
get_info(exp, pattern, type)
}
\arguments{
\item{exp}{object of class \code{experiment}.}

\item{pattern}{can be either \code{Parameters} or \code{Outputs} for
parameters and observation rates, respectively.}

\item{type}{can be either \code{name} or \code{type} for parameters and
observation rates' names and parameter types respectively.}
}
\value{
A vector of parameters or variables names, or a vector of parameters
types.
}
\description{
Get information on the model defined in the GAML file linked to an
\code{experiment} object.
}
\examples{
file <- system.file("models", "sir.gaml", package = "rama")
exp <- load_experiment("sir", file)
get_info(exp, "Parameters", "name")
get_info(exp, "Parameters", "type")
get_info(exp, "Outputs", "name")
get_info(exp, ".attrs", "experiment")
}
