% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_parameters_names}
\alias{get_parameters_names}
\alias{get_variables_names}
\alias{get_all_names}
\title{Get variables and parameters names}
\usage{
get_parameters_names(file)

get_variables_names(file)

get_all_names(file)
}
\arguments{
\item{file}{Name of a GAML file.}
}
\value{
A vector of parameters and / or variables names.
}
\description{
Retrieves the names of the parameters and variables of model in a GAML file.
}
\details{
The GAML file is parsed directly from R and does not requires to launch java
and GAMA. Thus, much faster!
}
\examples{
file <- system.file("models", "sir.gaml", package = "rama")
get_parameters_names(file)
get_variables_names(file)
get_all_names(file)
}
