% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{get_wkdir}
\alias{get_wkdir}
\alias{get_wkdir.default}
\alias{get_wkdir.experiment}
\title{Get and set the output directory}
\usage{
get_wkdir(x)

\method{get_wkdir}{default}(x)

\method{get_wkdir}{experiment}(x)
}
\arguments{
\item{x}{An object of class \code{experiment}.}
}
\value{
The path of the directory where the simulations outputs will be saved.
}
\description{
These functions allow to get and set the path to the directory where the
simulations outputs will be saved.
}
\examples{
exp1 <- load_experiment("sir", system.file("examples", "sir.gaml", package = "rama"))
get_wkdir(exp1)

}
