% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{getmodelparameter}
\alias{getmodelparameter}
\title{Get parameters of a gama model}
\usage{
getmodelparameter(modelfile, experimentname)
}
\arguments{
\item{modelfile}{relative or absolute path pointing on your model file}

\item{experimentname}{name of the loaded experiment}
}
\description{
Load experiment meta-data from a model file and an experimentname
}
\keyword{internal}
