\name{hiv}
\docType{data}
\alias{hiv}
\title{Cellular gene expression upon human immunodeficiency virus type 1
  infection of CD4+-T-Cell lines}
\description{
  This data set consists of 4 experiments using the same RNA preparation on 4
  different slides. The expression levels of ~7000 cellular RNA transcripts were assessed
  in CD4-T-cell lines at time $t=24$ hour after infection with HIV virus
  type 1. The first 4 columns correspond to the first treatment state (hiv
  infected). The second four represent the control state. The experiment
  is a balanced dye swap experiment. Finally, the last two columns
  contain the row and column positions of each gene on the array (slide).
}
\usage{data(hiv)}
\source{http://expression.microslu.washington.edu/expression/vantwoutjvi2002.html}
\references{
  van't Wout, A. B. ,Lehrma, G. K., Mikheeva, S. A., O'Keeffe, G. C.,
  Katze, M. G., Bumgarner, R. E., Geiss, G. K. and Mullins, J. I.
  Cellular gene expression upon human immunodeficiency virus type 1 infection of CD4+-T-Cell lines
  Journal of Virology, 2003. 77(2):1392-1402.}
\keyword{datasets}
