% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{is.experiment}
\alias{is.experiment}
\title{Test if experiment}
\usage{
is.experiment(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
The function returns `TRUE` or `FALSE` depending on whether its
argument is of chatacter type or not
}
\description{
Tests for objects of type \code{"experiment"}.
}
\examples{
gaml_file <- system.file("examples", "sir.gaml", package = "rama")
exp1 <- load_experiment("sir", gaml_file, "sir")

is.experiment(exp1)
}
