% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{map_experiment}
\alias{map_experiment}
\title{Create an experiment object using a data frame and an experiment object
as template}
\usage{
map_experiment(df, exp, dir = "")
}
\arguments{
\item{df}{A data frame used to initialize an experiment object.}

\item{exp}{An experiment object used as template.}

\item{dir}{Either absolute path or name output of directory. In the latter
case, current directory will be used. If \code{dir} is not
specified, name of model will be used to create an output
directory in the current directory.}
}
\description{
Create an experiment object using a data frame and an experiment object
as template
}
\examples{
exp1 <- load_experiment("sir",
    system.file("examples", "sir.gaml", package = "rama"))
df <- data.frame(matrix(1, nrow=5, ncol=12))
exp2 <- map_experiment(df, exp1)

}
