% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{observation}
\alias{observation}
\title{Extract monitoring rates}
\usage{
observation(x)
}
\arguments{
\item{x}{An object of class \code{experiment}.}
}
\value{
A data frame that is a subset of the inputed \code{experiment} object.
}
\description{
Subsets the columns of an \code{experiment} object that correspond to the
observation rates of the monitored variables.
}
\examples{
exp1 <- load_experiment("sir", system.file("examples", "sir.gaml", package = "rama"), "sir")
exp2 <- repl(exp1, 10)
observation(exp2)

}
