% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_dir.R
\name{output_dir}
\alias{output_dir}
\alias{output_dir.default}
\alias{output_dir.experiment}
\title{Get the output directory}
\usage{
output_dir(exp)

\method{output_dir}{default}(exp)

\method{output_dir}{experiment}(exp)
}
\arguments{
\item{exp}{An object of class \code{experiment}.}
}
\value{
The path of the directory where the simulations outputs will be
saved.
}
\description{
These functions allow to get and set the path to the directory where the
simulations outputs will be saved.
}
\examples{
exp1 <- load_experiment("sir", system.file("examples", "sir.gaml",
                        package = "rama"))
output_dir(exp1)

}
