% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realExp.R
\name{realexp}
\alias{realexp}
\title{Restore the correct values of an output}
\usage{
realexp(output, exp0)
}
\arguments{
\item{output}{An object of class \code{list}.}

\item{exp}{An object of class \code{experiment}.}
}
\value{
Returns a list of dataframes, one for each experiment.
}
\description{
It puts a NA whenever an observation was in fact not made in GAMA.
}
\details{
From the list of frequencies of each simulation of the experiment
the values incorrectly associated to the last value observed are reset to NA.
}
\examples{
exp0 <- experiment(
data.frame(S0 = c(900, 100), # this is a data frame of 2 lines
          I0 = c(100, 900),
          R0 = 0,
          beta = 2,
          gamma = .15),
data.frame(S = 1, I = 2, R = 3),
tmax = 1000,
seed = 1,
model = gaml_file,
experiment = "sir"
)
exp1 <- load_experiment("sir", gaml_file, "sir")
exp0 <- repl(exp1,3)
exp0[1,"p_S0"] <- 500
exp0[2,"p_S0"] <- 800
exp0[3,"p_S0"] <- 200
exp0[1,"p_I0"] <- 500
exp0[2,"p_I0"] <- 200
exp0[3,"p_I0"] <- 800
exp0[2,"r_S"] <- 2
exp0[3,"r_S"] <- 3
exp0[3,"r_I"] <- 8

otp <- run_experiment(exp0)
str(opt)
newoutput <- realexp(otp,exp0)
str(newoutput)

}
