% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{repl}
\alias{repl}
\alias{repl.default}
\alias{repl.experiment}
\title{Replicate an experiment}
\usage{
repl(x, n)

\method{repl}{default}(x, n)

\method{repl}{experiment}(x, n)
}
\arguments{
\item{x}{An object of class \code{experiment}.}

\item{n}{The number of times to repeat the experiment \code{x}.}
}
\value{
An object of class \code{experiment} that is the replication of the
inputed \code{x} argument.
}
\description{
Produces an \code{experiment} object from the replication of an
\code{experiment} object.
}
