% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Run an experiment}
\usage{
run_experiment(exp, hpc = 1, save = FALSE, path = NULL,
  force = FALSE, display = FALSE, append = TRUE)
}
\arguments{
\item{exp}{an XML file containing the experiment.}

\item{hpc}{number of threads used by GAMA to run the experiment.}

\item{save}{save the outputs to disk or not, default = FALSE.}

\item{path}{directory to save the outputs, default = NULL.
If `save = TRUE` and `path` is not
specified, current working directory is used to save the outputs.}

\item{force}{overwrite the specified `path` or not, default = FALSE.}

\item{display}{output images are saved or not, default = FALSE.}

\item{append}{append outputs to experiment, default = TRUE. It is not possible
to set both `add_exp` and `save` as `FALSE`.}
}
\description{
From a `experiment` object, run an experiment by creating an XML file of the
experiment (\code{\link[rama]{save_to_gama}}) and by calling gama
(\code{\link[rama]{call_gama}}) and returns a list of data frame, one by
simulation.
}
\examples{
#load experiment
gaml_file <- system.file("models", "sir.gaml", package = "rama")
exp1 <- load_experiment("sir", gaml_file)
# run experiment
out <- run_experiment(exp1)

# for a more complexe example
exp1 <- repl(exp1, 2)
exp1$r_R <- 2:3
output <- run_experiment(exp1)
str(output)
}
