% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{save_to_gama.experiment}
\alias{save_to_gama.experiment}
\title{Save an experiment plan to gama xml file}
\usage{
\method{save_to_gama}{experiment}(plan, file = "out.xml")
}
\arguments{
\item{plan}{Object of class experiment containing all experiment to do}

\item{file}{The path where you want to save the file.}
}
\description{
Save an xml \code{file} containing the experiment plan defined in the object of class \code{experiment}.
}
