% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_model.R
\name{set_model}
\alias{set_model}
\title{This function allows to change the model path of an experiment object}
\usage{
set_model(model, exp)
}
\arguments{
\item{model}{Path of new gaml model file}

\item{exp}{Experiment in question}
}
\description{
This function allows to change the model path of an experiment object
}
\examples{
exp1 <- load_experiment("sir", system.file("examples", "sir.gaml",
                        package = "rama"))
model(exp1)
set_model(exp1, system.file("examples", "CopyOfsir.gaml",
                        package = "rama"))
model(exp1)

}
