% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{show_experiment}
\alias{show_experiment}
\title{List the experiments of a model and their types}
\usage{
show_experiment(file)
}
\arguments{
\item{file}{path to a gaml model file.}
}
\description{
List the experiments of a given model.
}
