% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startexpplan.R
\name{writemodelparameterfile}
\alias{writemodelparameterfile}
\title{store experimentplan}
\usage{
writemodelparameterfile(experimentplan, outfile = "")
}
\arguments{
\item{experimentplan}{experiment plan to store}
}
\description{
Save an experiment plan to an xml file. it returns the path of the xml file
}
