% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_dir.R
\name{create_output_dir}
\alias{create_output_dir}
\title{Create output folder}
\usage{
create_output_dir(exp, dir = "")
}
\arguments{
\item{exp}{an object of class `experiment`}

\item{dir}{name of directory to saved the output of gama in the experiment
output directory. If not specified, experiment name will be used.}
}
\description{
Creates output folder named by the `experiment` name. If multiple
`experiment` are called with the same name, a numeric is append to avoid
confusion.
}
\examples{
#load experiment
gaml_file <- system.file("examples", "sir.gaml", package = "rama")
exp1 <- load_experiment("sir", gaml_file, "sir")

#create path to folder in working direction
output_dir <- create_output_dir(exp1, "")
parameter_xml_file <-  paste0(name(exp1), ".xml")
}
