% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{defpath}
\alias{defpath}
\title{This function allows to config Gama path and Java heap size to run Gama as
global options.}
\usage{
defpath(path, Xmx = "2048m", Xms = "512m")
}
\arguments{
\item{path}{Path to Gama}

\item{Xmx}{Maximum heap size}

\item{Xms}{Initial heap size}
}
\description{
This function allows to config Gama path and Java heap size to run Gama as
global options.
}
\examples{
defpath(path = "/Applications/Gama.app/", Xmx = "4096m", Xms = "512m")

}
