% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{experiment}
\alias{experiment}
\title{Create an experiment}
\usage{
experiment(parameters, obsrates, tmax, seed, model, experiment)
}
\arguments{
\item{parameters}{Either a data frame containing the paramters values, or a
character vector containing the names of the columns of the
first-position data frame argument that correspond to the
parameters.}

\item{obsrates}{Either a data frame containing the monitoring rates of the
observed variables, or a character vector containing the
names of the columns of the first-position data frame
argument that correspond to the monitored variables.}

\item{tmax}{xxx}

\item{seed}{xxx}

\item{model}{xxx}

\item{experiment}{xxx}

\item{...}{One or two data frames.}
}
\description{
Creates an object of class \code{experiment}.
}
\details{
This function is polymorph and can takes input of different kind to achieve
this. TO BE DESCRIBED IN MORE DETAIL.
}
\examples{
my_experiment <- experiment(
  expand.grid(S0 = c(900, 950, 999),
              I0 = c(100, 50, 1),
              R0 = 0,
              beta = 1.5,
              gamma = .15),
  data.frame(S = 1, I = 1, R = 1),
  tmax = 1000,
  seed = 1,
  model = "/Users/choisy/Dropbox/aaa/r-and-gama/rama/inst/examples/sir.gaml",
  experiment = "sir"
)
# If we want to change the seeds:
my_experiment$seed <- 1:9
}
