% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R, R/is.experiment.R
\name{experiment}
\alias{experiment}
\alias{experiment.default}
\alias{experiment.character}
\alias{experiment.numeric}
\alias{experiment.data.frame}
\alias{is.experiment}
\title{Create an object of class \code{experiment}}
\usage{
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
=======
>>>>>>> Fix experiment and documentation
=======
>>>>>>> conflict resolved
experiment(parameters, obsrates, tmax, seed, experiment, model, dir, df)
=======
experiment(parameters, obsrates, tmax, seed, experiment, model, dir, ...)
>>>>>>> Fix experiment and documentation
<<<<<<< HEAD
<<<<<<< HEAD
=======
experiment(parameters, obsrates, tmax, seed, experiment, model, dir, df)
>>>>>>> consistency of generic/method of experiment
=======
>>>>>>> Fix experiment and documentation
=======
=======
experiment(parameters, obsrates, tmax, seed, experiment, model, dir, df)
>>>>>>> conflict resolved
>>>>>>> conflict resolved

\method{experiment}{default}(parameters, obsrates, tmax, seed, experiment,
  model, dir, df)

\method{experiment}{character}(parameters = NULL, obsrates = NULL,
  tmax = NULL, seed = NULL, experiment = NULL, model = NULL,
  dir = "", df = NULL)

\method{experiment}{numeric}(parameters = NULL, obsrates = NULL,
  tmax = NULL, seed = NULL, experiment = NULL, model = NULL,
  dir = "", df = NULL)

\method{experiment}{data.frame}(parameters = NULL, obsrates = NULL,
  tmax = NULL, seed = NULL, experiment = NULL, model = NULL,
  dir = "", ...)

is.experiment(exp)
}
\arguments{
\item{parameters}{Vector of column names or indexes in the \code{df} that
will be used as parameters in the experiment.}

\item{obsrates}{Vector of column names or indexes in the \code{df} that will
be used as obs_rates rats in the experiment.}

\item{tmax}{Name or index of the column in the \code{df} that will be
used as final step in the experiment.}

\item{seed}{Name or index of the column in the \code{df} that will be
used as seed in the experiment.}

\item{experiment}{name to the model linked to the experiment}

\item{model}{path to the model linked to the experiment}

\item{dir}{Either absolute path or name output of directory. In the latter
case, current directory will be used. If \code{dir} is not
specified, name of model will be used to create an output
directory in the current directory.}

\item{df}{A data frame used to initialize an experiment object.}

\item{...}{Additional paramaters}

\item{exp}{object to be tested}

\item{...}{Additional paramaters}
}
\value{
The function returns `TRUE` or `FALSE` depending on whether its
argument is of chatacter type or not
}
\description{
Tests for objects of type \code{"experiment"}.
}
\examples{
# Experiment constructor with a dataframe and indexes/names of columns
# indicating parameters, observation rates, tmax and seed

df <- data.frame("S0" = rep(999, 5), "I0" = rep(1, 5), "R0" = rep(0, 5),
                "beta" = rep(1.5, 5), "gama" = runif (5, 0, 1),
                "S" = rep(1, 5), "I" = rep(1, 5), "R" = rep(1, 5),
                "a" = rep(1000, 5), "b" = rep(1, 5))
exp1 <- experiment(parameters = c("S0", "I0", "R0", "beta", "gama"),
                  obsrates = c("S", "I", "R"),
                  tmax = "a",
                  seed = "b",
                  experiment = "sir",
                  model =
                    system.file("examples", "sir.gaml", package = "rama"),
                  df = df)
exp2 <- experiment(parameters = c(1:5),
                  obsrates = c(6:8),
                  tmax = 9,
                  seed = 10,
                  experiment = "sir",
                  model =
                    system.file("examples", "sir.gaml", package = "rama"),
                  dir = "my_sir_model",
                  df = df)

# Experiment constructor that uses for data frames (of paramaters,
# observation rates, tmax and seed) as input.

df1 <- data.frame("S0" = rep(999, 5), "I0" = rep(1, 5), "R0" = rep(0, 5),
                 "beta" = rep(1.5, 5), "gama" = runif (5, 0, 1))
df2 <- data.frame("S" = rep(1, 5), "I" = rep(1, 5), "R" = rep(1, 5))
tmax <- rep(1000, 5)
seed <- rep(1, 5)
exp3 <- experiment(parameters = df1,
                   obsrates = df2,
                   tmax = tmax,
                   seed = seed,
                   experiment = "sir",
                   model =
                   system.file("examples", "sir.gaml", package = "rama"))

is.experiment(exp)
}
