% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R, R/is.experiment.R
\name{experiment}
\alias{experiment}
\alias{experiment.default}
\alias{experiment.data.frame}
\alias{is.experiment}
\title{Create an object of class \code{experiment}}
\usage{
experiment(df, parameters = NULL, obsrates = NULL, tmax = NULL,
  seed = NULL, experiment = NULL, model = NULL, dir = "")

\method{experiment}{default}(df, parameters = NULL, obsrates = NULL,
  tmax = NULL, seed = NULL, experiment = NULL, model = NULL,
  dir = "")

\method{experiment}{data.frame}(df, parameters = NULL, obsrates = NULL,
  tmax = NULL, seed = NULL, experiment = NULL, model = NULL,
  dir = "")

is.experiment(exp)
}
\arguments{
\item{df}{A data frame}

\item{parameters}{Vector of column names or indexes in the \code{df} that will be
used as parameters in the experiment.}

\item{obsrates}{Vector of column names or indexes in the \code{df} that will be
used as obs_rates rats in the experiment.}

\item{tmax}{Name or index of the column in the \code{df} that will be
used as final step in the experiment.}

\item{seed}{Name or index of the column in the \code{df} that will be
used as seed in the experiment.}

\item{experiment}{name to the model linked to the experiment}

\item{model}{path to the model linked to the experiment}

\item{dir}{Name of the output directory to be created in the current directory.
If not specified, name of the model will be used}

\item{exp}{object to be tested}
}
\value{
The function returns `TRUE` or `FALSE` depending on whether its
argument is of chatacter type or not
}
\description{
Tests for objects of type \code{"experiment"}.
}
\examples{
df <- data.frame("S0" = rep(999, 5), "I0" = rep(1, 5), "R0" = rep(0, 5),
                "beta" = rep(1.5, 5), "gama" = runif (5, 0, 1),
                "S" = rep(1, 5), "I" = rep(1, 5), "R" = rep(1, 5),
                "a" = rep(1000, 5), "b" = rep(1, 5))
exp <- experiment(df,
                  parameters = c("S0", "I0", "R0", "beta", "gama"),
                  obsrates = c("S", "I", "R"),
                  tmax = "a",
                  seed = "b",
                  experiment = "sir",
                  model = system.file("examples", "sir.gaml", package = "rama"))
exp <- experiment(df,
                  parameters = c(1:5),
                  obsrates = c(6:8),
                  tmax = 9,
                  seed = 10,
                  experiment = "sir",
                  model = system.file("examples", "sir.gaml", package = "rama"),
                  dir = "my_sir_model")

is.experiment(exp)
}
