% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R, R/is.experiment.R
\name{experiment}
\alias{experiment}
\alias{experiment.default}
\alias{experiment.character}
\alias{experiment.numeric}
\alias{experiment.data.frame}
\alias{experiment.experiment}
\alias{is.experiment}
\title{Create an object of class \code{experiment}}
\usage{
experiment(df, parameters, ...)

\method{experiment}{default}(df, parameters = NULL, obsrates = NULL,
  tmax = NULL, seed = NULL, experiment = NULL, model = NULL,
  dir = "")

\method{experiment}{character}(df, parameters = NULL, obsrates = NULL,
  tmax = NULL, seed = NULL, experiment = NULL, model = NULL,
  dir = "")

\method{experiment}{numeric}(df, parameters = NULL, obsrates = NULL,
  tmax = NULL, seed = NULL, experiment = NULL, model = NULL,
  dir = "")

\method{experiment}{data.frame}(parameters = NULL, obsrates = NULL,
  tmax = NULL, seed = NULL, experiment = NULL, model = NULL,
  dir = "")

\method{experiment}{experiment}(df, exp, dir = "")

is.experiment(exp)
}
\arguments{
\item{df}{A data frame}

\item{parameters}{Vector of column names or indexes in the \code{df} that will be
used as parameters in the experiment.}

\item{...}{Additional paramaters}

\item{obsrates}{Vector of column names or indexes in the \code{df} that will be
used as obs_rates rats in the experiment.}

\item{tmax}{Name or index of the column in the \code{df} that will be
used as final step in the experiment.}

\item{seed}{Name or index of the column in the \code{df} that will be
used as seed in the experiment.}

\item{experiment}{name to the model linked to the experiment}

\item{model}{path to the model linked to the experiment}

\item{dir}{Name of the output directory to be created in the current directory.
If not specified, name of the model will be used}

\item{exp}{object to be tested}
}
\value{
The function returns `TRUE` or `FALSE` depending on whether its
argument is of chatacter type or not
}
\description{
Tests for objects of type \code{"experiment"}.
}
\examples{
# Experiment constructor with a dataframe and indexes/names of columns
# indicating parameters, observation rates, tmax and seed

df <- data.frame("S0" = rep(999, 5), "I0" = rep(1, 5), "R0" = rep(0, 5),
                "beta" = rep(1.5, 5), "gama" = runif (5, 0, 1),
                "S" = rep(1, 5), "I" = rep(1, 5), "R" = rep(1, 5),
                "a" = rep(1000, 5), "b" = rep(1, 5))
exp1 <- experiment(df,
                  parameters = c("S0", "I0", "R0", "beta", "gama"),
                  obsrates = c("S", "I", "R"),
                  tmax = "a",
                  seed = "b",
                  experiment = "sir",
                  model = system.file("examples", "sir.gaml", package = "rama"))
exp2 <- experiment(df,
                  parameters = c(1:5),
                  obsrates = c(6:8),
                  tmax = 9,
                  seed = 10,
                  experiment = "sir",
                  model = system.file("examples", "sir.gaml", package = "rama"),
                  dir = "my_sir_model")

# Experiment constructor that uses for data frames (of paramaters, observation
# rates, tmax and seed) as input.

df1 <- data.frame("S0" = rep(999, 5), "I0" = rep(1, 5), "R0" = rep(0, 5),
                 "beta" = rep(1.5, 5), "gama" = runif (5, 0, 1))
df2 <- data.frame("S" = rep(1, 5), "I" = rep(1, 5), "R" = rep(1, 5))
tmax <- rep(1000, 5)
seed <- rep(1, 5)
exp3 <- experiment(parameters = df1, obsrates = df2,
                   tmax = tmax, seed = seed,
                   experiment = "sir",
                   model = system.file("examples", "sir.gaml", package = "rama"))

# Experiment constructor that uses a data frame and an experiment as template
# to create an experiment object.

df4 <- data.frame(matrix(1, nrow=5, ncol=12))
exp4 <- experiment(df4, exp1)
is.experiment(exp)
}
