% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_info}
\alias{get_info}
\title{Get information on the model}
\usage{
get_info(exp, pattern, type)
}
\arguments{
\item{exp}{Experiment object}

\item{pattern}{can be either 'Parameters' or 'Outputs' for parameters and
observation rates, respectively.}

\item{type}{can be 'name' for parameter and observation rate names, '
type' for parameter types.}
}
\value{
A vector of parameter or variable names,
}
\description{
Get information on the model
}
\examples{
file <- system.file("models", "sir.gaml", package = "rama")
exp <- load_experiment("sir", file)
get_info(exp, "Parameters", "name")
get_info(exp, "Parameters", "type")
get_info(exp, "Outputs", "name")
get_info(exp, ".attrs", "experiment")
}
