% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{list_experiment}
\alias{list_experiment}
\title{List the experiments of a model}
\usage{
list_experiment(x)
}
\arguments{
\item{x}{path to a gaml model file.}
}
\description{
List the experiments of a given model.
}
