% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{load_experiment}
\alias{load_experiment}
\title{Load An Experiment}
\usage{
load_experiment(experiment, model)
}
\arguments{
\item{experiment}{The name of the experiment to load.}

\item{model}{The name of the file from which to load the experiment.}
}
\description{
Loads an experiment from a model specified in a \code{gaml} file and returns
an object of class \code{plan}.
}
