% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{load_experiment}
\alias{load_experiment}
\title{Load An Experiment}
\usage{
load_experiment(experiment, model, dir = "")
}
\arguments{
\item{experiment}{The name of the experiment to load.}

\item{model}{The name of the file from which to load the experiment.}

\item{dir}{The name of the directory to save the output of the runs for each
model. If not specified, name of the gaml file will be used}
}
\description{
Loads an experiment from a model specified in a \code{gaml} file and returns
an object of class \code{experiment}.
}
\examples{
# Looking at the \\code{sir.gaml} file in the \\code{examples} directory of the
# \\code{rama} library:
gaml_file <- system.file("examples", "sir.gaml", package = "rama")

# Loading the experiment \\code{sir} from this gaml file:
exp1 <- load_experiment("sir", gaml_file)

# Checking the class:
class(exp1)

}
