% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_experiment.R
\name{load_experiment}
\alias{load_experiment}
\title{Load an experiment from a model}
\usage{
load_experiment(exp, model, dir = "")
}
\arguments{
\item{exp}{The name of the experiment to load.}

\item{model}{The name of the file from which to load the experiment.}

\item{dir}{The name of the directory to save the output of the runs for each
model. If not specified, name of the gaml file will be used}
}
\description{
Loads an experiment from a model specified in a \code{.gaml} file and returns
an object of class \code{experiment}.
}
\details{
The \code{rama} package contains a library of model in the \code{examples}
directory of the \code{rama} package file hierarchy. These models can be
accessed with the \code{system.file()} function as explained in the example.
}
\examples{
# Looking at the \\code{sir.gaml} file in the \\code{examples} directory of the
# \\code{rama} library:
# Loading an experiment:
exp1 <- load_experiment("sir", system.file("examples", "sir.gaml", package = "rama"))

# Checking the class:
class(exp1)

}
