% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_experiment.R
\name{load_experiment}
\alias{load_experiment}
\title{Load an experiment from a GAML file}
\usage{
load_experiment(exp, model)
}
\arguments{
\item{exp}{The name of the experiment to load.}

\item{model}{The name of the GAML file from which to load the experiment.}
}
\description{
Loads an experiment from a \code{.gaml} file and returns an object of class
\code{experiment}.
}
\details{
The \code{rama} package contains an internal collection of GAMA models. These
models are specified, at minima, by a \code{.gaml}. Additional files such as
shapefile can be used to specify a model, in which case they are in the same
directory as the \code{.gaml} file. The internal collection of GAMA models is
in the \code{examples} directory of the \code{rama} package file hierarchy.
These models can be accessed with the \code{\link[base]{system.file}}
function as explained in the example.
}
\examples{
# Listing the models available in the "examples" directory of the "rama" library:
grep(".gaml", dir(system.file("models", package = "rama")), value = TRUE)
# Loading experiment "sir" from the "sir.gaml" file:
exp1 <- load_experiment("sir",
                        system.file("models", "sir.gaml", package = "rama"))
# Checking the class:
class(exp1)
}
