% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model<-}
\alias{model<-}
\title{Set the model of an experiment}
\usage{
model(exp) <- value
}
\arguments{
\item{exp}{Experiment in question}

\item{value}{Path of new gaml model file}
}
\description{
This function allows to change the model path of an experiment object
}
\examples{
exp1 <- load_experiment("sir", system.file("models", "sir.gaml",
                        package = "rama"))
model(exp1)
model(exp1) <-  system.file("models", "CopyOfsir.gaml",
                        package = "rama")
model(exp1)
}
