% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\alias{model.default}
\alias{model.experiment}
\title{Get the model of an experiment}
\usage{
model(exp)

\method{model}{default}(exp)

\method{model}{experiment}(exp)
}
\arguments{
\item{exp}{An object of class \code{experiment}.}
}
\value{
The path to the \code{.gaml} file that contains the model definition.
}
\description{
These functions allow to get and set the path to the \code{.gaml} file that
contains the model of an \code{experiment} object.
}
\examples{
exp1 <- load_experiment("sir", system.file("models", "sir.gaml",
                        package = "rama"))
model(exp1)
model(exp1) <-  system.file("models", "CopyOfsir.gaml",
                        package = "rama")
model(exp1)
}
