% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realExp.R
\name{realexp}
\alias{realexp}
\title{Restore the correct values of an output}
\usage{
realexp(output, exp0)
}
\arguments{
\item{output}{An object of class \code{list}.}

\item{exp0}{An object of class \code{experiment}.}
}
\value{
Returns a list of dataframes, one for each experiment.
}
\description{
It puts a NA whenever an observation was in fact not made in GAMA.
}
\details{
From the list of frequencies of each simulation of the experiment
the values incorrectly associated to the last value observed are reset to NA.
}
\examples{
df <-   data.frame(S0 = c(900, 800, 500), # this is a data frame of 3 lines
                  I0 = c(100, 200, 500),
                  R0 = 0,
                  beta = 1.5,
                  gamma = .15,
                  S = c(1,2,3),
                  I = c(2,4,6),
                  R =c(10,20,30),
                  nbiter = 1000,
                  seed = "123456789")

exp0 <- experiment(
 df,
 parameters = c("S0","I0","R0","beta","gamma"),
 obsrates  = c("S", "I", "R"),
 tmax = "nbiter",
 seed = "seed",
 experiment = "sir",
 model = system.file("examples", "sir.gaml", package = "rama"),
 dir = "testsir"
)


exp0
otp <- run_experiment(exp0)
str(otp)
newoutput <- realexp(otp,exp0)
str(newoutput)

}
