% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Run an experiment}
\usage{
run_experiment(exp, hpc = 1)
}
\arguments{
\item{exp}{an XML file containing the experiment}

\item{hpc}{numeric}
}
\description{
From a `experiment` object, run an experiment by creating an XML file of the
experiment (\code{\link[rama]{save_to_gama}}) and by calling gama
(\code{\link[rama]{call_gama}}) and returns a list of data frame, one by
simulation.
}
\examples{
#load experiment
gaml_file <- system.file("models", "sir.gaml", package = "rama")
exp1 <- load_experiment("sir", gaml_file)
# run experiment
out <- run_experiment(exp1)

# for a more complexe example
exp1 <- repl(exp1, 2)
exp1$r_R <- 2:3
output <- run_experiment(exp1)
str(output)
}
