% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Run an experiment}
\usage{
run_experiment(exp, hpc = 1, output_dir = "",
  parameter_xml_file = "")
}
\arguments{
\item{exp}{an XML file containing the experiment}

\item{hpc}{numeric}

\item{output_dir}{path to saved the output of gama}

\item{parameter_xml_file}{path to folder containing the xml file

#' @examples
#load experiment
gaml_file <- system.file("examples", "sir.gaml", package = "rama")
exp1 <- load_experiment("sir", gaml_file, "sir")
# run experiment
out <- run_experiment(exp1)}
}
\description{
From a `experiment` object, run an experiment by creating an XML file of the
experiment (\code{\link[rama]{save_to_gama}}) and by calling gama
(\code{\link[rama]{call_gama}}) and returns a list of data frame, one by
simulation and also the simulation output of Gama in XML file stored in
`output_dir` (created by \code{\link[rama]{create_output_dir}})
}
