% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_gama.R
\name{save_to_gama}
\alias{save_to_gama}
\alias{save_to_gama.experiment}
\title{Save an experiment plan to a GAMA XML file}
\usage{
save_to_gama(exp, file)

\method{save_to_gama}{experiment}(exp, file = "out.xml")
}
\arguments{
\item{exp}{An object of class \code{experiment}.}

\item{file}{The path to the output XML file.}
}
\description{
Save an object of class \code{experiment} to an XML file GAMA-compliant.
}
\examples{
#load experiment
gaml_file <- system.file("examples", "sir.gaml", package = "rama")
exp1 <- load_experiment("sir", gaml_file, "sir")

save_to_gama(exp1)

}
