% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_functions.R
\name{save_to_gama}
\alias{save_to_gama}
\title{Save an experiment plan to a GAMA XML file}
\usage{
save_to_gama(plan, file)
}
\arguments{
\item{plan}{Object of class \code{experiment}.}

\item{file}{The path to the output XML file.}
}
\description{
Save an object of class \code{experiment} to an XML file GAMA-compliant.
}
