% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_gama.R
\name{save_to_gama}
\alias{save_to_gama}
\alias{save_to_gama.default}
\alias{save_to_gama.experiment}
\title{Save an experiment plan to a GAMA XML file}
\usage{
save_to_gama(exp, parameter_xml_file)

\method{save_to_gama}{default}(exp, parameter_xml_file)

\method{save_to_gama}{experiment}(exp, parameter_xml_file = "")
}
\arguments{
\item{exp}{An object of class \code{experiment}.}

\item{parameter_xml_file}{name of XML parameter file. This file is created
in the working directory of `exp`. If not
specified, name of `exp`` is used.}
}
\description{
Save an object of class \code{experiment} to an XML file GAMA-compliant.
}
\examples{
gaml_file <- system.file("models", "sir.gaml", package = "rama")
exp1 <- load_experiment("sir", gaml_file, "sir")
save_to_gama(exp1)
}
