% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_experiment.R
\name{show_experiment}
\alias{show_experiment}
\title{List the experiments of a model and their types}
\usage{
show_experiment(file)
}
\arguments{
\item{file}{Path to a \code{.gaml} file.}
}
\description{
List the experiments of a given model.
}
\examples{
gaml_file <- system.file("examples", "sir.gaml", package = "rama")
show_experiment(gaml_file)

}
